requirements.txt
# all imported files from main.py,mail.py,audio.py,app.py for a requirment.txt file 

os==2.0.0
pyaudio==0.2.11
speechrecognition==3.8.1
pyttsx3==2.90
os-sys==2.0.0
bardapi==0.3.6
time==2.0.0
dotenv==0.0.5
google-api-python-client==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4